changed extension metadata update to work on initialization
