+++ packages/page-settings/src/index.tsx
@@ -21,7 +21,6 @@ export { useCounter };
 function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const { api, isApiConnected, isApiReady, isDevelopment } = useApi();
-  const numExtensions = useCounter();
 
   const items = useMemo(() => [
     {
@@ -30,7 +29,6 @@ function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<P
       text: t('General')
     },
     {
-      count: numExtensions,
       name: 'metadata',
       text: t('Metadata')
     },
@@ -42,7 +40,7 @@ function SettingsApp ({ basePath, onStatusChange }: Props): React.ReactElement<P
       name: 'i18n',
       text: t('Translate')
     }
-  ], [numExtensions, t]);
+  ], [t]);
 
   const hidden = useMemo(
     () => (isApiConnected && isApiReady)
